fix: change signatures' paramater types to double from float

+++ SolarWatch/SolarWatch/Services/CoordAndDateProcessor.cs
@@ -6,19 +6,19 @@ namespace SolarWatch.Services;
 
 public class CoordAndDateProcessor : ICoordAndDateProcessor
 { 
-    public async Task<string> GetSunriseTime(float lat, float lon, string date)
+    public async Task<string> GetSunriseTime(double lat, double lon, string date)
     {
         var (sunrise, _) = await GetSunRiseSetTime(lat, lon, date);
         return sunrise;
     }
     
-    public async Task<string> GetSunsetTime(float lat, float lon, string date)
+    public async Task<string> GetSunsetTime(double lat, double lon, string date)
     {
         var (_, sunset) = await GetSunRiseSetTime(lat, lon, date);
         return sunset;
     }
 
-    private async Task<(string, string)> GetSunRiseSetTime(float lat, float lon, string date)
+    private async Task<(string, string)> GetSunRiseSetTime(double lat, double lon, string date)
     {
         var url = $"https://api.sunrise-sunset.org/json?lat={lat}&lng={lon}&date={date}";
         using (var client = new HttpClient())
